♻️ [#63] LocationManager 리팩토링
