+++ DoriDori_iOS/DoriDori_iOS/Scene/Question/QuestionType.swift
@@ -1,0 +1,13 @@
+//
+//  QuestionType.swift
+//  DoriDori_iOS
+//
+//  Created by Seori on 2022/08/23.
+//
+
+import Foundation
+
+enum QuestionType {
+    case user(userID: UserID)
+    case community
+}
